==> Data_Analysis_Pandas/ch_04/2-marek.ipynb <==
import numpy as np
import pandas as pd

weather = pd.read_csv('data/nyc_weather_2018.csv', parse_dates=['date'])
weather.head()

fb = pd.read_csv('data/fb_2018.csv', index_col='date', parse_dates=['date'])
fb.head()

fb.assign(
    Z_score= lambda x: x.volume.sub(x.volume.mean()).div(x.volume.std()).abs()
).query('Z_score > 3')

fb.assign(
    volume_pct_change = fb.volume.pct_change(),
    pct_rank = lambda x: x.volume_pct_change.abs().rank(ascending=False)
).nsmallest(5,'pct_rank')

(fb.volume.value_counts() > 1).any()

from window_calc import window_calc

window_calc??
