==> Data_Analysis_Pandas/ch_04/3.2- marek.ipynb <==
token ='jyJkyAhgZxfZTifXwMPichXZAgoZTcrv'

    
import requests

def make_request(endpoint, payload=None):
    return requests.get(
    f"https://www.ncdc.noaa.gov/cdo-web/api/v2/{endpoint}",
    headers={
        'token':'jyJkyAhgZxfZTifXwMPichXZAgoZTcrv'
    },
    params=payload
    )


response = make_request('datasets')
requests.get?

# get NYC daily summaries data 
response = make_request(
    'data', 
    {
        'datasetid' : 'GHCND',
        'stationid' : 'GHCND:USW00094728',
        'locationid' : 'CITY:US360019',
        'startdate' : '2018-10-01',
        'enddate' : '2018-10-31',
        'datatypeid' : ['TMIN', 'TMAX', 'TOBS'], # temperature at time of observation, min, and max
        'units' : 'metric',
        'limit' : 1000
    }
)
response.status_code

def get_item(name, what, endpoint, start=1, end=None):
    
    mid = (start + (end if end else 1)) // 2
    name = name.lower()
    
    payload = {
        'datasetid' : 'GHCND',
        'sortfield' : 'name',
        'offset' : mid, # we will change the offset each time
        'limit' : 1 # we only want one value back
    }
    
    response = make_request(endpoint, {**payload, **what})
    
    if response.ok:
        end = end if end else response.json()['metadata']['resultset']['count']
        print(f'start: {start} mid: {mid} end: {end}')
        
        current_name = response.json()['results'][0]['name'].lower()
        print(current_name)
        
        if name in current_name:
            return response.json()['results'][0]
        else:
            if start >= end:
                return {}
            elif name < current_name:
                return get_item(name, what, endpoint, start, mid-1)
            elif name > current_name:
                return get_item(name, what, endpoint, mid+1, end)
            
    else:
        print(f'Response not ok, status {response.status_code}')
    

def get_location(name):
    
    return get_item(name, {'locationcategoryid':'CITY'}, 'locations')

## Begining of testing phase

payload = {
        'datasetid' : 'GHCND',
        'sortfield' : 'name',
        'offset' : 1255, # we will change the offset each time
        'limit' : 1
        }

response1 = make_request('locations', {**payload, 'locationcategoryid':'CITY'}) ## how does it work - SOLVED - nested dictionary

response1.json()['results'][0]['name']

payload = {
        'datasetid' : 'GHCND',
        'sortfield' : 'name',
        'offset' : 1255, # we will change the offset each time
        'limit' : 1
        }

def rand_print(**kwargs):
    print(kwargs)
    
rand_print(**payload)

## The end of testing phase

nyc = get_location('new york')
nyc

nyc['id']

central_park = get_item('NY City Central Park', {'locationid' : nyc['id']}, 'stations')

central_park

# get NYC daily summaries data 
response = make_request(
    'data', 
    {
        'datasetid' : 'GHCND',
        'stationid' : 'GHCND:USW00014732',
        'locationid' : nyc['id'],
        'startdate' : '2018-10-01',
        'enddate' : '2018-10-31',
        'datatypeid' : ['TMIN', 'TMAX', 'TAVG'], # temperature at time of observation, min, and max
        'units' : 'metric',
        'limit' : 1000
    }
)
response.status_code

response.json()

df = pd.DataFrame(response.json()['results'])
df.head()

df.to_csv('marek_df.csv', index=False)

df.datatype.value_counts()
